+++ tests/features/test_node_ignore.py
@@ -5,7 +5,9 @@ from ididi.errors import UnsolvableDependencyError
 
 
 class IgnoreNode:
-    def __init__(self, name: str, age: int): ...
+    def __init__(self, name: str, age: int):
+        self.name=name
+        self.age=age
 
 
 def test_direct_resolve_fail():
@@ -27,3 +29,14 @@ def test_resolve_with_ignore():
     dg = DependencyGraph()
     dg.node(ignore=("name", int))(IgnoreNode)
     dg.static_resolve(IgnoreNode)
+
+
+def test_resolve_without_ignore():
+    dg = DependencyGraph()
+    dg.node(ignore=("name", int))(IgnoreNode)
+    with pytest.raises(TypeError):
+        dg.resolve(IgnoreNode)
+
+    n = dg.resolve(IgnoreNode, name="test", age=3)
+    assert n.name == "test"
+    assert n.age == 3